# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Hexsatisfaction",
        "title": "Hexsatisfaction API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/purchase/api/file/{fileID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find purchases by file id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "summary": "FindByFileID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File id",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Purchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagError"
                        }
                    },
                    "404": {
                        "description": "No purchases",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagEmptyError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagError"
                        }
                    }
                }
            }
        },
...
        "/purchase/api/user/{userID}/file/{fileID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Find purchases by user id and file id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purchase"
                ],
                "summary": "FindByUserIDAndFileID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File id",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Purchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagError"
                        }
                    },
                    "404": {
                        "description": "No purchases",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagEmptyError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.SwagError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middleware.SwagEmptyError": {
            "type": "object"
        },
        "middleware.SwagError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Purchase": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "fileID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}